backend-api/requirements.txt
Django>=4.2,<5
djangorestframework>=3.14
django-cors-headers>=4.3
dj-database-url>=2.2
psycopg2-binary>=2.9
python-dotenv>=1.0